gladieter/requirements.txt
amqp==1.4.9
amqpconsumer==1.5
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.24
dj-database-url==0.4.1
Django==1.9.8
django-allauth==0.28.0
django-autofixture==0.12.1
django-rest-auth==0.8.2
djangorestframework==3.5.3
dnspython==1.15.0
gunicorn==19.6.0
honcho==0.7.1
kombu==3.0.37
librabbitmq==1.6.1
oauthlib==2.0.0
pika==0.10.0
psycopg2==2.6.2
python-dateutil==2.6.0
python-openid==2.2.5
pytz==2016.10
PyXB==1.2.5
requests==2.11.1
requests-oauthlib==0.7.0
six==1.10.0
whitenoise==3.2
